+++ tests/test_utils.py
@@ -16,45 +16,11 @@ from click import echo
 import pytest
 import types
 
-#===============================================================================
-# Make sure that the current directory is the project directory.
-# 'make test" and 'pytest' are generally run from the project directory.
-# However, if we run/debug this file in eclipse, we end up in test
-if os.getcwd().endswith('tests'):
-    echo(f"Changing current working directory"
-         f"\n  from '{os.getcwd()}'"
-         f"\n  to   '{os.path.abspath(os.path.join(os.getcwd(),'..'))}'\n")
-    os.chdir('..')
-#===============================================================================    
-# Make sure that we can import the module being tested. When running 
-# 'make test" and 'pytest' in the project directory, the current working
-# directory is not automatically added to sys.path.
-if not ('.' in sys.path or os.getcwd() in sys.path):
-    p = os.path.abspath('.')
-    echo(f"Adding '{p}' to sys.path.\n")
-    sys.path.insert(0, p)
-echo(f"sys.path = \n{sys.path}".replace(',','\n,'))
-#===============================================================================
 
 from tests.helpers import report, in_empty_tmp_dir 
 from et_micc import cli,commands
 import et_micc.utils
 import et_micc.logging_tools
-
-
-#===============================================================================
-# test scenario blocks
-#===============================================================================
-def run(runner,arguments,input_='short description'):
-    """
-    create a project 
-    """
-    result = runner.invoke( cli.main
-                          , arguments
-                          , input=input_
-                          )
-    report(result)
-
 
 #===============================================================================
 #   Tests
@@ -117,14 +83,19 @@ def test_get_project_path():
     p = Path.cwd()
     p = et_micc.utils.get_project_path(p)
     assert p==(Path.home() / 'software/dev/workspace/et-micc')
-    
+
+
+def test_get_dependencies():
+    deps = et_micc.utils.get_dependencies()
+    print(deps)
+    assert 'numpy' in deps
 
 # ==============================================================================
 # The code below is for debugging a particular test in eclipse/pydev.
 # (normally all tests are run with pytest)
 # ==============================================================================
 if __name__ == "__main__":
-    the_test_you_want_to_debug = test_get_project_path # test_scenario_1
+    the_test_you_want_to_debug = test_get_dependencies
 
     print(f"__main__ running {the_test_you_want_to_debug}")
     the_test_you_want_to_debug()